+++ src/app/Fake.DotNet.FSFormatting/FSFormatting.fs
@@ -1,22 +1,27 @@
 ﻿/// Contains tasks which allow to run FSharp.Formatting for generating documentation.
-[<System.Obsolete "use Fake.DotNet.FSFormatting instead">]
-module Fake.FSharpFormatting
+module Fake.DotNet.FSFormatting
 
-#nowarn "44"
+open System
+open System.Diagnostics
+open System.IO
+open Fake.Core
+open Fake.Core.String
+open Fake.Core.Globbing
+open Fake.Core.Process
+open Fake.IO.FileSystem
+open Fake.IO.FileSystem.Operators
+
 /// Specifies the fsformatting executable
-[<System.Obsolete "use Fake.DotNet.FSFormatting instead">]
-let mutable toolPath = 
-    findToolInSubPath "fsformatting.exe" (currentDirectory @@ "tools" @@ "FSharp.Formatting.CommandTool" @@ "tools")
+let mutable toolPath =
+    Tools.findToolInSubPath "fsformatting.exe" (Directory.GetCurrentDirectory() @@ "tools" @@ "FSharp.Formatting.CommandTool" @@ "tools")
 
 /// Runs fsformatting.exe with the given command in the given repository directory.
-[<System.Obsolete "use Fake.DotNet.FSFormatting instead">]
 let run command = 
     if 0 <> ExecProcess (fun info -> 
                 info.FileName <- toolPath
                 info.Arguments <- command) System.TimeSpan.MaxValue
     then failwithf "FSharp.Formatting %s failed." command
-    
-[<System.Obsolete "use Fake.DotNet.FSFormatting instead">]
+
 let CreateDocs source outputDir template projectParameters = 
     let command = 
         projectParameters
@@ -31,8 +36,7 @@ let CreateDocs source outputDir template projectParameters =
         |> separated " "
     run command
     printfn "Successfully generated docs for %s" source
-    
-[<System.Obsolete "use Fake.DotNet.FSFormatting instead">]
+
 let CreateDocsForDlls outputDir templatesDir projectParameters sourceRepo dllFiles = 
     for file in dllFiles do
         projectParameters
@@ -40,7 +44,7 @@ let CreateDocsForDlls outputDir templatesDir projectParameters sourceRepo dllFil
         |> Seq.concat
         |> Seq.append 
                 ([ "metadataformat"; "--generate"; "--outdir"; outputDir; "--layoutroots"; "./help/templates/"; 
-                    templatesDir; "--sourceRepo"; sourceRepo; "--sourceFolder"; currentDirectory; "--parameters" ])
+                    templatesDir; "--sourceRepo"; sourceRepo; "--sourceFolder"; Directory.GetCurrentDirectory(); "--parameters" ])
         |> Seq.map (fun s -> 
                 if s.StartsWith "\"" then s
                 else sprintf "\"%s\"" s)